feat: New disabled mixin
